Line mode added to cmucam2 project
git-svn-id: https://cc3.svn.sourceforge.net/svnroot/cc3/trunk@155 111db651-790d-0410-a044-b0abe44a5820

+++ lib/cc3_ilp/cc3_color_track.h
@@ -15,7 +15,7 @@ typedef struct {
     uint32_t num_pixels;
     uint32_t int_density;
     uint8_t noise_filter;
-    uint32_t binary_scanline[MAX_BINARY_WIDTH];
+    uint8_t binary_scanline[MAX_BINARY_WIDTH];
     cc3_pixel_t upper_bound;
     cc3_pixel_t lower_bound;
     uint16_t scratch_x,scratch_y;
